refactor: refactor and modularize extension and LSP settings UI
- Refactor extension management by moving logic into standalone files: `extensionActions.kt`, `ExtensionActionButton.kt`, `ExtensionAuthorIcon.kt`, `MarkdownViewer.kt`, `ReviewsPage.kt`, and `SourceCodeSheet.kt`.
- Extract `LspInstanceCard` into a separate component to improve maintainability in LSP server detail screens.
- Introduce `StateScreen` component for consistent display of loading, error, and empty states.
- Add `timeAgo` utility in `Utils.kt` for relative timestamp formatting.
- Update UI icons to use `painterResource(drawables.open_in_new)` for external links in support and extension screens.
- Clean up `ExtensionScreen.kt`, `ExtensionDetail.kt`, and `LspServerDetail.kt` by removing redundant inline code and logic.
